Updated work on fixing window Selector
Technically, the window selector program will not be a part of the final
program. However, it will allow me to get practice working with hWnds,
treads, objects, logic, rendering and handling information
I need to move over the render thread, after I decide how I want to
implement it, and learn more about it. Then bug fixing/rendering to a
panel properly (had issues with it in prev. version)